Added the Neighbourhood table
It is a bunch of functions that need to be moved into their own file.
I could replace it with data, but I'm not sure that would be better...
We will see. Too many states to represent it elegantly as data.

Also added water, which erodes walls, and pools which have a chance of
sticking around.